requirements.txt
PyYAML
GitPython
numpy
